Begin work on stripping Authorization header where possible.

+++ Cardboard.Net.Core/Entities/Users/IUserDecoration.cs
@@ -6,6 +6,6 @@ public interface IUserDecoration : IMisskeyEntity
     int Angle { get; }
     bool FlipH { get; }
     Uri Url { get; }
-    int XOffset { get; }
-    int YOffset { get; }
+    float XOffset { get; }
+    float YOffset { get; }
 }
